backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.4
python-multipart==0.0.9
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.4.2
beautifulsoup4==4.12.3
httpx==0.27.0
redis==5.0.4
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
